Refactor it to avoid iterating the events multiple times
